Streamlined decompressDatafile function
No longer write to a temporary file on disk to  read back from it. Created a class that acts like a file object but is much quicker to access
Made file paths much more streamlined
Ideally we shouldn't be writing all the temporary files to disk as this is slow. Perhaps come back to this in the future