The start of a plugin system to allow multiple export options
Currently the only way to export data is by double clicking on an entry in the file tree that calls a function which points to the relevant code for that file type. This only allows one exporting function per entry in the file tree which is very limiting.
This should allow an endless amount and make things easier to develop