moved the texture related code to the more appropriate places
Split up the texture code
The code called from the UI is in the right_click_methods folder. This is just enough to point to the other functions
The code in type_readers is the code that reads the raw file and returns data
The bulk of the exporting code has been moved to misc as it can be used in multiple games. This could be implemented in right_click_methods