[Internal] Turned part of the dds export into a property
The DDS class now creates a string of the dds file as a property. This means it can be used elsewhere without saving to a file.
Need to try and work out how to solve the DX10 issue. The exe being used only works on saved files not binary streams (also it isn't cross platform)