[Internals] Mesh cleanup and rework
In the stored model format the vertices start from 0 as is common in programming.
The OBJ format starts from 1 so an offset was added. This offset has been moved to the OBJ exporter from the model reading where it was before.
Hidden model variables in privite attributes and implemented getters for them.
Fixed some issues with model reading. It looks like some modes use a block based system and others don't. This fix should work most of the time but isn't how it is supposed to be implemented.