[Internals] Swapped out the old light dictionary for a numpy array
The old system had the light dictionary stored as a json file which was slow to load adding a couple of seconds to the load time. The new numpy method is a lot quicker to load and save but might be negligably slower while running. It is not used that much though so it should not be an issue