[Internals] Replaced numpy unique with a set
There needs to be some way to check if a file is in the light dictionary. This was done by appending it anyway and getting the unique of the list but that is slow.
There is now a set built off the file ids and forge files which is used to check if it is present in the array before appending