[Internals] Modified tempfiles again
Changed tempfiles to be somewhere between the old and new systems. Searching for values in the numpy array was slow so this is implemented as a dictionary. Both systems are maintained while the program is running but it only uses the dictionaries. When the program is closed the new files are appened to the numpy array and it is saved out.
Also means we only merge the numpy array once.