[UI] Fixed some incorrect overwriting issues
In the file/directory select tool "" is returned if the cancel button is pressed rather than None as was expected. This meant the path was being changed if the cancel button was pressed.
The newly implemented menus were not checking if the cancel button had been pressed so were overwriting data when not intended.