[UI] Added icons and dynamic based on the theme
Added icons to the file tree to better show what the file is. They are loaded from each theme and fallback to the default ones.
To add resource icons add an image Qt can read to the icons folder with the file name being the capitalised big endian file type.